+++ src/Board.ts
@@ -1,0 +1,64 @@
+import './moves.css'
+import { toDests } from './utils'
+import { getPieceLocations } from './ChessMaths'
+import m from 'mithril'
+import { Chessground } from 'chessground'
+
+export const Board = cell =>
+  m(
+    '.board-wrapper',
+    {},
+    m('#board.board', {
+      oncreate: vnode => {
+        console.log(
+          'Board oncreate',
+          cell.state,
+          cell.state.chess.history({ verbose: true })
+        )
+        let lastMove = cell.state.chess.history({ verbose: true }).pop()
+        let ground = Chessground(vnode.dom, {
+          fen: cell.state.fen,
+          turnColor: cell.state.chess.turn() == 'w' ? 'white' : 'black',
+          lastMove: lastMove ? [lastMove.from, lastMove.to] : undefined,
+          movable: {
+            free: false,
+            dests: toDests(cell.state.chess),
+            color: cell.state.chess.turn() == 'w' ? 'white' : 'black',
+          },
+          highlight: {
+            check: true,
+          },
+          events: {
+            move: (orig, dest) => {
+              let state = cell.getState()
+              console.log('move', orig, dest)
+              state.chess.move({ from: orig, to: dest })
+              // console.log(cell.state.chess.fen())
+              ground.set({
+                // fen: cell.state.chess.fen(),
+                movable: {
+                  color: state.chess.turn() == 'w' ? 'white' : 'black',
+                  dests: toDests(state.chess),
+                },
+                check: state.chess.inCheck(),
+              })
+              console.log('move', orig, dest)
+              cell.update({
+                fen: state.chess.fen(),
+                pgn: state.chess.pgn(),
+                moves: state.chess
+                  .history({ verbose: true })
+                  .map(move => move.san),
+                // ground: ground,
+              })
+            },
+          },
+        })
+        window.ground = ground
+        cell.update({
+          fen: cell.state.chess.fen(),
+          pgn: cell.state.chess.pgn(),
+        })
+      },
+    })
+  )
