+++ src/State.ts
@@ -1,0 +1,20 @@
+import { Chess } from 'chess.js'
+import { Chessground } from 'chessground'
+
+export interface State {
+  page: string
+  login?: {
+    username: string
+    password: string
+  }
+  data?: string[] | string
+  chess: Chess
+  ground?: typeof Chessground
+  fen?: string
+  moves?: string[]
+  pgn?: string
+  isHistoryCollapsed: boolean
+  isFENCollapsed: boolean
+  isPGNCollapsed: boolean
+  isInfoCollapsed: boolean
+}
