+++ src/index.html
@@ -13,5 +13,5 @@
     <div id="app"></div>
     <div id="tracer" class="component"></div>
   </body>
-  <script type="module" src="meiosis.ts"></script>
+  <script type="module" src="index.ts"></script>
 </html>
