+++ src/meiosis.css
@@ -15,11 +15,6 @@
 
 #game.component {
     display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    width: 100vw;
-    height: 80vh;
-    position: relative;
 }
 
 .stacked.component {
@@ -50,7 +45,6 @@
 }
 
 .info {
-    min-width: 20vw;
 }
 
 .history {
