+++ src/moves.css
@@ -1,3 +1,3 @@
 td.move-number, td.move {
-    padding: 0 10px;
+    padding-right: 2rem;
 }
